cleanup(src/utils): Remove keg util methods   * Remove unused code   * Remove runYarnScript && runRepoScript   * Minor cleanup items
