+++ src/utils/getters/getHubRepos.js
@@ -1,9 +1,10 @@
 const path = require('path')
+const { getRepoPath } = require('./getRepoPath')
 const { asyncCmd } = require('@keg-hub/spawn-cmd')
-const { getRepoPath } = require('../getters/getRepoPath')
 const { generalError } = require('../error/generalError')
 const { fileSys, Logger } = require('@keg-hub/cli-utils')
 const { isFunc, pickKeys, eitherArr } = require('@keg-hub/jsutils')
+
 
 const { readFileSync } = fileSys
