--- src/utils/helpers/index.js
@@ -1,4 +1,0 @@
-module.exports = {
-  ...require('./runYarnScript'),
-  ...require('./updateLocationContext'),
-}
