--- src/utils/hub/index.js
@@ -1,4 +1,0 @@
-module.exports = {
-  ...require('./getHubRepos'),
-  ...require('./runRepoScript'),
-}
