+++ src/utils/index.js
@@ -6,7 +6,6 @@ module.exports = {
   ...require('./globalConfig'),
   ...require('./getters'),
   ...require('./helpers'),
-  ...require('./hub'),
   ...require('./process'),
   ...require('./log'),
   ...require('./network'),
